refactor: create test_utils module with feature developer-mode for generator testing methods
